Day 12 part 1 fix some compiler warnings define the equality operator for Pos (need to define both <=> and == to deduce all comparison operators)

+++ src/Day06/Day06.cpp
@@ -38,7 +38,7 @@ int main()
 		{
 			if (line[col] == '#')
 			{
-				obstacles.emplace(col, row);
+				obstacles.emplace( static_cast<int>(col), static_cast<int>(row) );
 			}
 			else if (line[col] == '^')
 			{
